requirements.txt
requests
beautifulsoup4
pillow
moviepy
gTTS
instagrapi
transformers
gpt4all
